713 moved subclass properties
