+++ pykechain/models/property_datetime.py
@@ -1,16 +1,13 @@
 import datetime
-import warnings
-
+from typing import Union, Text
 from pykechain.exceptions import IllegalArgumentError
-from pykechain.models.property import Property
+from pykechain.models import Property
+from pykechain.models.input_checks import check_datetime
 from pykechain.utils import parse_datetime
 
 
-class DatetimeProperty(Property):  # pragma: no cover
-    """A virtual object representing a KE-chain reference property."""
-
-    def __index__(self, json, **kwargs):
-        super(DatetimeProperty, self).__init__(json, **kwargs)
+class DatetimeProperty(Property):
+    """A virtual object representing a KE-chain datetime property."""
 
     @property
     def value(self):
@@ -24,16 +21,15 @@ class DatetimeProperty(Property):  # pragma: no cover
 
     @value.setter
     def value(self, value):
-        if isinstance(value, datetime.datetime):
-            if not value.tzinfo:
-                warnings.warn("The value '{}' is naive and not timezone aware, use pytz.timezone info. "
-                              "This date is interpreted as UTC time.".format(value.isoformat(sep=' ')))
-
-            self._put_value(value.isoformat(sep='T'))
+        if value is None:
+            self._put_value(None)
+        elif isinstance(value, datetime.datetime):
+            self._put_value(check_datetime(dt=value, key='value'))
         else:
             raise IllegalArgumentError('value should be a datetime.datetime() object')
 
     def to_datetime(self):
+        # type: () -> Union[type(None), datetime.datetime]
         """Retrieve the data value of a property.
 
         Setting this value will immediately update the property in KE-chain.
@@ -41,3 +37,21 @@ class DatetimeProperty(Property):  # pragma: no cover
         :returns: the value
         """
         return parse_datetime(self._value)
+
+    @staticmethod
+    def to_iso_format(date_time):
+        # type: (datetime.datetime) -> Text
+        """Convert a datetime object to isoformat."""
+        return date_time.isoformat()
+
+    def serialize_value(self, value) -> str:
+        """
+        Serialize the value to be set on the property by checking for datetime objects.
+
+        :param value: non-serialized value
+        :type value: Any
+        :return: serialized value
+        """
+        if isinstance(value, datetime.datetime):
+            value = self.to_iso_format(value)
+        return value
