+++ pykechain/models/property_multi_reference.py
@@ -1,97 +1,58 @@
-from six import text_type, string_types
+from typing import List, Optional, Text, Union, Any, Dict
 
+from pykechain.enums import Category, FilterType
+from pykechain.models.base_reference import _ReferencePropertyInScope
 from pykechain.models.part import Part
-from pykechain.models.property import Property
-from pykechain.utils import is_uuid
+from pykechain.models.widgets.helpers import _check_prefilters, _check_excluded_propmodels
 
 
-class MultiReferenceProperty(Property):  # pragma: no cover
+class MultiReferenceProperty(_ReferencePropertyInScope):
     """A virtual object representing a KE-chain multi-references property.
 
     .. versionadded:: 1.14
     """
 
-    def __init__(self, json, **kwargs):
-        """Construct a MultiReferenceProperty from a json object."""
-        super(MultiReferenceProperty, self).__init__(json, **kwargs)
+    REFERENCED_CLASS = Part
 
-        self._cached_values = None
-
-    @property
-    def value(self):
-        """Value of a reference property.
-
-        You can set the reference with a Part, Part id or None value.
-        Ensure that the model of the provided part, matches the configured model
-
-        :return: a :class:`Part` or None
-        :raises APIError: When unable to find the associated :class:`Part`
-
-        Example
-        -------
-        Get the wheel reference property
-
-        >>> part = project.part('Bike')
-        >>> wheels_ref_property = part.property('Wheels')
-        >>> isinstance(wheels_ref_property, MultiReferenceProperty)
-        True
-
-        The value returns a list of Parts or is an empty list
-
-        >>> type(wheels_ref_property.value) in (list, tuple)
-        True
-
-        Get the selection of wheel instances:
-
-        >>> wheel_choices = wheels_ref_property.choices()
-
-        Choose random wheel from the wheel_choices:
-
-        >>> from random import choice
-        >>> wheel_choice_1 = choice(wheel_choices)
-        >>> wheel_choice_2 = choice(wheel_choices)
-
-        Set chosen wheel
-        1: provide a single wheel:
-
-        >>> wheels_ref_property.value = [wheel_choice_1]
-
-        2: provide multiple wheels:
-
-        >>> wheels_ref_property.value = [wheel_choice_1, wheel_choice_2]
-
+    def _retrieve_objects(self, **kwargs) -> List[Part]:
         """
-        if not self._value:
-            return None
-        if not self._cached_values and isinstance(self._value, (list, tuple)):
-            ids = [v.get('id') for v in self._value]
-            self._cached_values = list(self._client.parts(id__in=','.join(ids), category=None))
-        return self._cached_values
+        Retrieve a list of Parts.
 
-    @value.setter
-    def value(self, value):
-        value_to_set = []
-        if isinstance(value, (list, tuple)):
-            for item in value:
-                if isinstance(item, Part):
-                    value_to_set.append(item.id)
-                elif isinstance(item, (string_types, text_type)) and is_uuid(item):
-                    # tested against a six.text_type (found in the requests' urllib3 package) for unicode
-                    # conversion in py27
-                    value_to_set.append(item)
+        :param kwargs: optional inputs
+        :return: list of Part objects
+        """
+        if self._value is None:
+            return []
+
+        ids = []
+        for value in self._value:
+            if isinstance(value, dict):
+                ids.append(value.get('id'))
+            elif isinstance(value, str):
+                ids.append(value)
+            else:
+                raise ValueError('Value "{}" must be a dict with field `id` or a UUID.'.format(value))
+
+        parts = []
+
+        if ids:
+            if self.category == Category.MODEL:
+                parts = [self._client.part(pk=ids[0], category=None)]
+            elif self.category == Category.INSTANCE:
+                # Retrieve the referenced model for low-permissions scripts to enable use of the `id__in` key
+                if False:  # TODO Check for script permissions in order to skip the model() retrieval
+                    models = [None]
                 else:
-                    raise ValueError("Reference must be a Part, Part id or None. type: {}".format(type(item)))
-        elif isinstance(value, type(None)):
-            # clear out the list
-            value_to_set = None
-        else:
-            raise ValueError(
-                "Reference must be a list (or tuple) of Part, Part id or None. type: {}".format(type(value)))
+                    models = self.model().value
+                if models:
+                    parts = list(self._client.parts(
+                        id__in=','.join(ids),
+                        model=models[0],
+                        category=None,
+                    ))
+        return parts
 
-        # we replace the current choices!
-        self._value = self._put_value(value_to_set)
-
-    def choices(self):
+    def choices(self) -> List[Part]:
         """Retrieve the parts that you can reference for this `MultiReferenceProperty`.
 
         This method makes 2 API calls: 1) to retrieve the referenced model, and 2) to retrieve the instances of
@@ -102,15 +63,100 @@ class MultiReferenceProperty(Property):  # pragma: no cover
 
         Example
         -------
-        >>> property = project.part('Bike').property('a_multi_reference_property')
-        >>> reference_part_choices = property.choices()
+        >>> reference_property = project.part('Bike').property('a_multi_reference_property')
+        >>> referenced_part_choices = reference_property.choices()
 
         """
-        # from the reference property (instance) we need to get the value of the reference property in the model
-        # in the reference property of the model the value is set to the ID of the model from which we can choose parts
-        model_parent_part = self.part.model()  # makes single part call
-        property_model = model_parent_part.property(self.name)
-        referenced_model = property_model.value and property_model.value[0]  # list of seleceted models is always 1
-        possible_choices = self._client.parts(model=referenced_model)  # makes multiple parts call
+        # Check whether the model of this reference property (possible itself) has a configured value
+        referenced_model = self.model()._value  # type: Optional[List[Dict]]
+        if referenced_model:
+            # If a model is configured, retrieve its ID
+            choices_model_id = referenced_model[0].get('id')
 
-        return possible_choices
+            # Retrieve all part instances with this model ID
+            possible_choices = self._client.parts(model_id=choices_model_id)  # makes multiple parts call
+
+            return possible_choices
+        else:
+            return list()
+
+    def set_prefilters(
+            self,
+            property_models: List[Union[Text, 'AnyProperty']],
+            values: List[Any],
+            filters_type: List[FilterType],
+            overwrite: Optional[bool] = False
+    ) -> None:
+        """
+        Set the pre-filters on a `MultiReferenceProperty`.
+
+        :param property_models: `list` of `Property` models (or their IDs) to set pre-filters on
+        :type property_models: list
+        :param values: `list` of values to pre-filter on, value has to match the property type.
+        :type values: list
+        :param filters_type: `list` of filter types per pre-fitler, one of :class:`enums.FilterType`,
+                defaults to `FilterType.CONTAINS`
+        :type filters_type: list
+        :param overwrite: determines whether the pre-filters should be over-written or not, defaults to False
+        :type overwrite: bool
+
+        :raises IllegalArgumentError: when the type of the input is provided incorrect.
+        """
+        if not overwrite:
+            initial_prefilters = self._options.get('prefilters', {})
+            prefilter_string = initial_prefilters.get('property_value', '')
+            list_of_prefilters = prefilter_string.split(',') if prefilter_string else []
+        else:
+            list_of_prefilters = list()
+
+        new_prefilters = {
+            'property_models': property_models,
+            'values': values,
+            'filters_type': filters_type,
+        }
+
+        list_of_prefilters.extend(_check_prefilters(
+            part_model=self.value[0],
+            prefilters=new_prefilters,
+        ))
+
+        # Only update the options if there are any prefilters to be set, or if the original filters have to overwritten
+        if list_of_prefilters or overwrite:
+            options_to_set = self._options
+
+            # Always create the entire prefilter json structure
+            options_to_set['prefilters'] = {
+                'property_value': ','.join(list_of_prefilters) if list_of_prefilters else ""
+            }
+
+            self.edit(options=options_to_set)
+
+    def set_excluded_propmodels(
+            self,
+            property_models: List[Union[Text, 'AnyProperty']],
+            overwrite: Optional[bool] = False,
+    ) -> None:
+        """
+        Exclude a list of properties from being visible in the part-shop and modal (pop-up) of the reference property.
+
+        :param property_models: `list` of Property models (or their IDs) to exclude.
+        :type property_models: list
+        :param overwrite: flag whether to overwrite existing (True) or append (False, default) to existing filter(s).
+        :type overwrite bool
+
+        :raises IllegalArgumentError
+        """
+        if not overwrite:
+            list_of_propmodels_excl = self._options.get('propmodels_excl', [])
+        else:
+            list_of_propmodels_excl = list()
+
+        list_of_propmodels_excl.extend(_check_excluded_propmodels(
+            part_model=self.value[0],
+            property_models=property_models,
+        ))
+
+        options_to_set = self._options
+        options_to_set['propmodels_excl'] = list_of_propmodels_excl
+
+        self.edit(options=options_to_set)
