+++ pykechain/models/property_selectlist.py
@@ -1,3 +1,7 @@
+from abc import abstractmethod
+from copy import deepcopy
+from typing import Any, Iterable
+
 from jsonschema import validate
 
 from pykechain.enums import Category
@@ -6,13 +10,14 @@ from pykechain.models.property import Property
 from pykechain.models.validators.validator_schemas import options_json_schema
 
 
-class SelectListProperty(Property):  # pragma: no cover
+class _SelectListProperty(Property):
     """A select list property that needs to update its options."""
 
     def __init__(self, json, **kwargs):
         """Construct a Property from a json object."""
-        super(SelectListProperty, self).__init__(json, **kwargs)
-        self._value_choices = self._json_data.get('options').get('value_choices')
+        super().__init__(json, **kwargs)
+
+        self._value_choices = json.get('value_options').get('value_choices')
 
     @property
     def value(self):
@@ -23,16 +28,27 @@ class SelectListProperty(Property):  # pragma: no cover
 
         :raises APIError: when unable to set the value or the value is not in the list of options
         """
-        return self._value
+        return deepcopy(self._value)
 
     @value.setter
-    def value(self, value):
-        if value not in self.options and value is not None:
-            raise APIError('The new value `{}` of the single select list property should be in the list of '
-                           'options `{}`'.format(value, self.options))
+    def value(self, value: Any):
+        if value is not None:
+            self._check_new_value(value=value)
+
         if self._put_value(value):
             self._value = value
             self._json_data['value'] = value
+
+    @abstractmethod
+    def _check_new_value(self, value: Any):
+        """
+        Abstract method to validate the new value.
+
+        :param value: any new value
+        :return: sanitized value
+        :raises APIError: when unable to set the value or the value is not in the list of options
+        """
+        pass
 
     @property
     def options(self):
@@ -63,7 +79,7 @@ class SelectListProperty(Property):  # pragma: no cover
         List of options of this property to select from.
 
         """
-        return self._value_choices
+        return deepcopy(self._value_choices)
 
     @options.setter
     def options(self, options_list):
@@ -78,25 +94,56 @@ class SelectListProperty(Property):  # pragma: no cover
 
         # check if the options are not already set and indeed different
         if not self._value_choices or set(options_list) != set(self._value_choices):
-            self._put_options(options_list=options_list)
+            self._put_value_options(options_list=options_list)
             self._value_choices = options_list
 
-    def _put_options(self, options_list):
+    def _put_value_options(self, options_list):
+        new_options = self._options.copy()  # make a full copy of the dict not to only link it and update dict in place
+        new_options.update({"value_choices": options_list})
+
+        self._put_options(new_options)
+
+    def _save_representations(self, representation_options):
+        new_options = self._options.copy()  # make a full copy of the dict not to only link it and update dict in place
+        new_options.update({'representations': representation_options})
+
+        self._put_options(new_options)
+
+    def _put_options(self, new_options):
         """Save the options to KE-chain.
 
         Makes a single API call.
 
-        :param options_list: list of options to set.
+        :param new_options: list of options to set.
         :raises APIError: when unable to update the options
         """
-        new_options = self._options.copy()  # make a full copy of the dict not to only link it and update dict in place
-        new_options.update({"value_choices": options_list})
         validate(new_options, options_json_schema)
 
         url = self._client._build_url('property', property_id=self.id)
-        response = self._client._request('PUT', url, json={'options': new_options})
+        response = self._client._request('PUT', url, json={'value_options': new_options})
 
         if response.status_code != 200:  # pragma: no cover
-            raise APIError("Could not update property value. Response: {}".format(str(response)))
+            raise APIError("Could not update options of Property {}".format(self), response=response)
         else:
             self._options = new_options  # save the new options as the options
+
+
+class SelectListProperty(_SelectListProperty):
+    """A virtual object representing a KE-chain single-select list property."""
+
+    def _check_new_value(self, value: Any):
+        if value not in self.options:
+            raise APIError('The new value "{}" of the Property should be in the list of options:\n{}'.format(
+                value, self.options))
+
+
+class MultiSelectListProperty(_SelectListProperty):
+    """A virtual object representing a KE-chain multi-select list property."""
+
+    def _check_new_value(self, value: Iterable[Any]):
+        if not isinstance(value, (list, tuple)):
+            raise IllegalArgumentError('The new values must be provided as a list or tuple, "{}" is not.'.format(value))
+
+        if not all(v in self.options for v in value):
+            raise APIError('The new values "{}" of the Property should be in the list of options:\n{}'.format(
+                value, self.options))
